use unsigned long/size_t to avoid index overflow in the case of large problems

+++ examples/cholesky/dw_cholesky.c
@@ -221,7 +221,7 @@ void initialize_system(float **A, unsigned dim, unsigned pinned)
 
 	if (pinned)
 	{
-		starpu_malloc_pinned_if_possible((void **)A, dim*dim*sizeof(float));
+		starpu_malloc_pinned_if_possible((void **)A, (size_t)dim*dim*sizeof(float));
 	} 
 	else {
 		*A = malloc(dim*dim*sizeof(float));
